Reorder things
